new updates

+++ src/project_huishoudboekje/database.py
@@ -34,7 +34,7 @@ def read_sql_table_cats(to_records=True, fill_nan_end_year=False):
     conn.close()
 
     if fill_nan_end_year:
-        df['end_year'] = df['end_year'].astype(str).replace('', '9999').replace('None', '9999')
+        df['end_year'] = df['end_year'].astype(str).replace('', '9999').replace('None', '9999').astype('int')
 
     if to_records:
         return df.drop(columns=['id']).sort_values(by=['grouplevel', 'category', 'begin_year']).to_dict('records')
